Fixed popular page to use hook/Search hook can be reused.
